requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
Django==1.11.1
django-redis==4.9.0
djangorestframework==3.8.2
Flask==1.0.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.19
pycryptodome==3.6.6
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
